+++ NEWS.md
@@ -1,3 +1,10 @@
+# dynCorr 1.1.0
+* Expanded functionality to incorporate dynamic correlation estimation for datasets with non-integer valued indep variable. 
+* A column is added to return value of bootstrapCI to give direct visual comparison of cor estimation and confidence interval estimation. 
+* Table data.summary is added to return values of both bootstrapCI and dynamicCorrelation to summarize heterogeneity of dataset. Three columns included in this table are: sample.size (number of observations used in the calculation), min.max.time (max common observation time), max.max.time (max individual observation time in the sample).
+* Three new variables are introduced to both dynamicCorrelation and bootstrapCI: points.length, points.by, and min.obs. See help page for more detail.
+
+
 # dynCorr 1.0.0
 
 * Added a `NEWS.md` file to track changes to the package.